019-Django---proje3__djangomania__page_ve_todo_list/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-autoslug==1.9.8
sqlparse==0.4.3
